created a single `set_git_url` function to be used across providers and placed it in `gitops.rs` file
